Add ClientRepository interface
This commit adds the 'ClientRepository' interface, which extends 'JpaRepository<Client, Long>'. This interface is used for interacting with the 'Client' entity in the Spring Data JPA repositxtends 'JpaRepository<Client, Long>'. This interface is used fentity in the Spring Data JPA rephe database.or interacting with the 'Client' entity in the Spring Data JPAelete) operations on client data  repository. It allows us to perform CRUD (Create, Read, Update, Delete) operations on client data in the database.